Added Android jetpack navigation dependency and adapter architecture.
