+++ app/src/main/java/hr/foi/air/visualbrickfinder/SplashScreenActivity.java
@@ -10,11 +10,15 @@ import android.view.WindowManager;
 
 public class SplashScreenActivity extends AppCompatActivity {
 
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
+        //Removing status bar
         requestWindowFeature(Window.FEATURE_NO_TITLE);
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
+
         setContentView(R.layout.activity_splash_screen);
         new Handler().postDelayed(() -> {
             Intent mainActivityIntent = new Intent(SplashScreenActivity.this, MainActivity.class);
