+++ app/src/main/res/layout/activity_main.xml
@@ -4,39 +4,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:background="@drawable/app_background">
+    tools:context=".MainActivity">
 
-
-    <TextView
-        android:id="@+id/activity_main_txt_tap"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/activity_main_imgbtn_take_photo"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="25dp"
-        android:fontFamily="@font/roboto_medium"
-        android:text="@string/tap_find"
-        android:textColor="@color/colorOpacityWhite"
-        android:textSize="25sp"
-        android:textStyle="bold" />
-
-    <RelativeLayout
-        android:id="@+id/activity_main_rellay_pulse"
-        android:layout_width="180dp"
-        android:layout_height="150dp"
-        android:layout_centerInParent="true"
-        android:background="@drawable/button_pulse" />
-
-    <ImageButton
-        android:id="@+id/activity_main_imgbtn_take_photo"
-        android:layout_width="180dp"
-        android:layout_height="150dp"
-        android:layout_centerInParent="true"
-        android:background="@drawable/circular_button"
-        android:contentDescription="@string/wienerberger_logo"
-        android:src="@drawable/logo" />
-
+    <fragment
+        android:id="@+id/nav_host_fragment"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/nav_graph"/>
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/activity_main_view_btmnav"
