+++ app/src/main/res/layout/fragment_homepage.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".HomepageFragment"
+    android:background="@drawable/app_background">
+
+    <TextView
+        android:id="@+id/activity_main_txt_tap"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_above="@+id/activity_main_imgbtn_take_photo"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="25dp"
+        android:fontFamily="@font/roboto_medium"
+        android:text="@string/tap_find"
+        android:textColor="@color/colorOpacityWhite"
+        android:textSize="25sp"
+        android:textStyle="bold" />
+
+    <RelativeLayout
+        android:id="@+id/activity_main_rellay_pulse"
+        android:layout_width="180dp"
+        android:layout_height="150dp"
+        android:layout_centerInParent="true"
+        android:background="@drawable/button_pulse" />
+
+    <ImageButton
+        android:id="@+id/activity_main_imgbtn_take_photo"
+        android:layout_width="180dp"
+        android:layout_height="150dp"
+        android:layout_centerInParent="true"
+        android:background="@drawable/circular_button"
+        android:contentDescription="@string/wienerberger_logo"
+        android:src="@drawable/logo" />
+
+</RelativeLayout>
